feat: new role dynamic_variables
This role allows to define the dynamic variables with custom Ansible
filters instead of relying on the inventory.

By default, the j2 variables are defined in the inventory which is
managed by the end-user. A user might manage the inventory in a
different way than the core roles of BlueBanquise (e.g., SCM vs. rpm).
Moreover, if the core should add some new j2 variables or update some
logic in existing variables, the end-user would need to be very careful
when upgrading to apply the new logic in the inventory files.

See the README.rst for further details.